requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
pydantic-settings==2.2.*
python-dotenv==1.0.*
PyJWT==2.8.*
azure-identity==1.16.*
azure-keyvault-secrets==4.8.*
